housekeeping: move all repository files into repositories folder
