+++ lib/domain/services/user_service.dart
@@ -1,4 +1,4 @@
-import '../../data/user_repository.dart';
+import '../../data/repositories/user_repository.dart';
 import '../models/user.dart';
 
 /// Service for user module.
